Teach unified server to load assets.
