+++ packages/unified-server/.env.development
@@ -1,0 +1,3 @@
+VITE_CONNECTION_SERVER_URL=http://localhost:3000/connection
+VITE_LANGUAGE_PACK=@breadboard-ai/shared-ui/strings/en_US
+VITE_ASSET_PACK=/icons/
