+++ packages/unified-server/.env.production
@@ -1,0 +1,5 @@
+# To connect the Visual Editor to a Connection Server, provide the Connection
+# Server URL here.
+VITE_CONNECTION_SERVER_URL=
+VITE_LANGUAGE_PACK=@breadboard-ai/shared-ui/strings/en_US
+VITE_ASSET_PACK=/icons/
