+++ packages/unified-server/.gitignore
@@ -1,2 +1,3 @@
 public
 index.html
+icons
