+++ packages/unified-server/package.json
@@ -103,6 +103,7 @@
     "@typescript-eslint/eslint-plugin": "^7.18.0",
     "@typescript-eslint/parser": "^7.18.0",
     "ava": "^5.2.0",
+    "dotenv": "^16.4.7",
     "typescript": "^5.7.3",
     "wireit": "^0.14.10"
   },
